feedback second pass remove duplicate
