+++ packages/insomnia-app/send-request/index.ts
@@ -4,11 +4,9 @@ import 'regenerator-runtime/runtime';
 import { database } from '../app/common/database';
 import { BaseModel, types as modelTypes } from '../app/models';
 import * as models from '../app/models';
-import { getBodyBuffer } from '../app/models/response';
 import { Settings } from '../app/models/settings';
 import { _actuallySend } from '../app/network/curl';
-import { send } from '../app/network/network';
-import * as plugins from '../app/plugins';
+import { sendAndTransform } from '../app/network/network';
 
 // The network layer uses settings from the settings model
 // We want to give consumers the ability to override certain settings
@@ -42,31 +40,6 @@ export async function getSendRequestCallbackMemDb(environmentId: string, memDB: 
 
   // Return callback helper to send requests
   return async function sendRequest(requestId: string) {
-    return sendAndTransform(requestId, environmentId);
+    return sendAndTransform(requestId, environmentId, _actuallySend);
   };
-}
-async function sendAndTransform(requestId: string, environmentId?: string) {
-  try {
-    plugins.ignorePlugin('insomnia-plugin-kong-declarative-config');
-    plugins.ignorePlugin('insomnia-plugin-kong-kubernetes-config');
-    plugins.ignorePlugin('insomnia-plugin-kong-portal');
-    const res = await send(requestId, environmentId, undefined, _actuallySend);
-    const headersObj: Record<string, string> = {};
-
-    for (const h of res.headers || []) {
-      const name = h.name || '';
-      headersObj[name.toLowerCase()] = h.value || '';
-    }
-
-    const bodyBuffer = await getBodyBuffer(res) as Buffer;
-    return {
-      status: res.statusCode,
-      statusMessage: res.statusMessage,
-      data: bodyBuffer ? bodyBuffer.toString('utf8') : undefined,
-      headers: headersObj,
-      responseTime: res.elapsedTime,
-    };
-  } finally {
-    plugins.clearIgnores();
-  }
 }
